Added minheight and minwidth support for dimensions
Additional layout on right-side of existing inner layout on default page
using % height combined with minheight.  Minimums can now also be
applied to middle sections, but outer layouts have to have minimums that
fit them for it to be observed (i.e. the style properties are set but
are not honoured by the browsers as the parents have a smaller height.