Data Panel
Updated data tab and data panel so that the dimensions of the item being
displayed by the data panel are set the first time the panel tries to
render them (which means the panel becomes the parent vm and dimension
inheritance comes from it).

Also included workflowcanvas control to get a basic canvas template in
and taking up the available space (which was what initiated the changes
above).